stream graph capturing - error handling

+++ src/main/native/src/solver_stat.h
@@ -7,14 +7,14 @@ namespace solver {
 struct SolverStat {
     SolverStat() = default;
 
-    long long startTime;
-    long long stopTime;
-    long long timeDelta;
+    __int64 startTime;
+    __int64 stopTime;
+    __int64 timeDelta;
     size_t size;
     size_t coefficientArity;
     size_t dimension;
-    long long accEvaluationTime;
-    long long accSolverTime;
+    __int64 accEvaluationTime;
+    __int64 accSolverTime;
     bool convergence;
     double error;
     double constraintDelta;
